+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Master
 ### Features
 - New option filter to attach lazily like `$(sel).on('click', filter, ...)`. Attaching is done for the whole set, not on each individual element.
 - New option otherClose to support extra closing buttons.
+- New content filter 'text' for plain text messages.
 - Gallery has new callbacks `beforeImage` and `afterImage`.
 - Gallery defaults are accessible with $.featherlightGallery.defaults